+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/Util.cs
@@ -8,9 +8,9 @@ namespace Frends.SFTP.UploadFiles.Definitions;
 /// </summary>
 internal static class Util
 {
-    internal static string CreateUniqueFileName()
+    internal static string CreateUniqueFileName(string fileExtension)
     {
-        return Path.ChangeExtension("frends_" + DateTime.Now.Ticks + Path.GetRandomFileName(), "8CO");
+        return Path.ChangeExtension("frends_" + DateTime.Now.Ticks + Path.GetRandomFileName(), fileExtension);
     }
 
     internal static bool FileMatchesMask(string filename, string mask)